Implement sum-type

+++ app/models/excel_file.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Boolean, Column, ForeignKey, Integer, String
+from sqlalchemy import Boolean, Column, ForeignKey, Integer, String, Float
 
 from app.database import Base
 
@@ -8,4 +8,4 @@ class ExcelFile(Base):
     id = Column(Integer, primary_key=True, index=True)
     category_id = Column(Integer)
     file_addr = Column(String)
-    sum = Column(Integer, default=None, nullable=True)
+    sum = Column(Float, default=None, nullable=True)
